hidefile: Fix incorrect typings definition structure
