--- hidefile/winattr.d.ts
@@ -1,1 +1,0 @@
-declare module 'winattr';
