+++ types/hidefile/tslint.json
@@ -1,0 +1,1 @@
+{ "extends": "dtslint/dt.json" }
